+++ src/main/java/com/dsh105/holoapi/api/Hologram.java
@@ -315,7 +315,7 @@ public class Hologram {
 
         WrappedDataWatcher dw = new WrappedDataWatcher();
         dw.watch(10, this.tags[index].replace("%name%", observer.getName()));
-        dw.watch(11, Byte.valueOf(1));
+        dw.watch(11, Byte.valueOf((byte) 1));
         dw.watch(12, Integer.valueOf(-1700000));
         horse.setDataWatcher(dw.getHandle());
